feat: show node labels in minimap fullscreen
