+++ src/browser/components/world/nodes/file/index.tsx
@@ -1,6 +1,6 @@
 import { Clone } from "@react-three/drei";
-import type { Group } from "three";
 import { Backdrop } from "@/browser/components/world/backdrop";
+import type { TNodeProps } from "@/browser/components/world/nodes/";
 
 export const File = ({
   instance,
@@ -10,18 +10,11 @@ export const File = ({
   nodeRef,
   isColliding,
   isBrowserEnvironment,
-}: {
-  label: string;
-  index: number;
-  model: Group;
-  nodeRef: TNodeRef;
-  instance: TNodeInstance;
-  isColliding: boolean;
-  isBrowserEnvironment: boolean;
-}) => {
+  isMinimapFullScreen,
+}: TNodeProps) => {
   return (
     // @ts-expect-error
-    <group position-y={13}>
+    <group position-y={isMinimapFullScreen ? 20 : 13}>
       <Clone
         ref={(el) => {
           if (!nodeRef?.current) return;
@@ -29,10 +22,17 @@ export const File = ({
         }}
         position={instance.position}
         object={model}
-        scale={27}
+        scale={isMinimapFullScreen ? 30 : 27}
       >
-        <Backdrop label={label} yPosition={2.68} fontSize={0.35} />
-        {isColliding ? (
+        <Backdrop
+          label={label}
+          fontSize={isMinimapFullScreen ? 0.8 : 0.35}
+          yPosition={isMinimapFullScreen ? 0.1 : 2.68}
+          frontOffset={isMinimapFullScreen ? -1.9 : undefined}
+          flatten={isMinimapFullScreen}
+          maxWidth={isMinimapFullScreen ? undefined : label.length + 15}
+        />
+        {isColliding && !isMinimapFullScreen ? (
           <Backdrop
             label={
               isBrowserEnvironment
@@ -41,7 +41,7 @@ export const File = ({
             }
             color="white"
             yPosition={0.8}
-            fontSize={isBrowserEnvironment ? 0.25 : 0.25}
+            fontSize={0.25}
             frontOffset={-1}
             maxWidth={isBrowserEnvironment ? 3 : 2.5}
           />
