+++ src/browser/components/world/nodes/index.tsx
@@ -1,3 +1,4 @@
+import type { Group } from "three";
 import { Folder } from "@/browser/components/world/nodes/folder";
 import { File } from "@/browser/components/world/nodes/file";
 import { useFolderModel } from "@/browser/hooks/useModels/folder";
@@ -6,8 +7,20 @@ import { useNodeMovement } from "@/browser/hooks/useNode/movement";
 import { useNodePlacement } from "@/browser/hooks/useNode/placement";
 import { useNodeCollision } from "@/browser/hooks/useNode/collision";
 
+export type TNodeProps = {
+  label: string;
+  index: number;
+  model: Group;
+  nodeRef: TNodeRef;
+  instance: TNodeInstance;
+  isColliding: boolean;
+  isBrowserEnvironment: boolean;
+  isMinimapFullScreen: boolean;
+};
+
 export const Nodes = () => {
-  const { nodes, boatRef, nodeRef, isBrowserEnvironment } = useNodePlacement();
+  const { nodes, boatRef, nodeRef, isBrowserEnvironment, isMinimapFullScreen } =
+    useNodePlacement();
 
   const fileModel = useFileModel(isBrowserEnvironment);
   const folderModel = useFolderModel(isBrowserEnvironment);
@@ -26,6 +39,7 @@ export const Nodes = () => {
           nodeRef,
           isColliding,
           isBrowserEnvironment,
+          isMinimapFullScreen,
           label: instance.data.name,
           model: isFile ? fileModel : folderModel,
           key: `${instance.data.type}-${instance.key}`,
